chore: reorganize cache utils under server/utils/cache and unify CacheMessage type
