+++ packages/sdk/server/utils/index.ts
@@ -1,6 +1,6 @@
 export * from "./archive";
-export * from "./auto-kv-cache-history";
-export * from "./cache";
+export * from "./cache/auto-history";
+export * from "./cache/paths";
 export * from "./checksum";
 export * from "./config-transform";
 export * from "./formatting";
